refactoring: new file for extension DynamicAction
