+++ src/xcode/ENA/ENA/Source/Scenes/AppInformation/AppInformationDynamicAction.swift
@@ -1,0 +1,38 @@
+//
+// Corona-Warn-App
+//
+// SAP SE and all other contributors
+// copyright owners license this file to you under the Apache
+// License, Version 2.0 (the "License"); you may not use this
+// file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+import Foundation
+//import UIKit
+
+extension DynamicAction {
+	static var safari: Self {
+		.execute { viewController in
+			LinkHelper.showWebPage(from: viewController, urlString: AppStrings.SafariView.targetURL)
+		}
+	}
+	
+	static func push(model: DynamicTableViewModel, separators: Bool = false, withTitle title: String) -> Self {
+		.execute { viewController in
+			let detailViewController = AppInformationDetailViewController()
+			detailViewController.title = title
+			detailViewController.dynamicTableViewModel = model
+			detailViewController.separatorStyle = separators ? .singleLine : .none
+			viewController.navigationController?.pushViewController(detailViewController, animated: true)
+		}
+	}
+}
